Fix crash in discardPurgeableFiles with proper selection syntax.
The basic crash is that this codepath was doing a query to the download
provider without properly quoting its selection arguments (which is
mandated by the download provider).

A secondary crash is that the DESTINATION column wasn't readable.

This fixes bug 1941125.

Tested by force-calling discardPurgeableFiles for each download
to /cache, locally adding a few debugging statements to be sure
that the codepath was getting executed, and doing a market download.