Show correct date for downloads in the statusbar
The date was not set properly in notifications when the text for
expanded view was created.

This fix sets the displayed time to the download time for each
individual file. The DownloadProvider recreates the notifications
several times but don't set the time before the expanded message
is created. The expanded message will therefore display the time
the notification was created and not the time a file was downloaded.
The displayed time will in most case be the time when the last file
was downloaded. This fix sets the displayed time to the downloaded
time for each individual file.