New functional test for download manager public API.
I refactored a lot of the logic from DownloadManagerFunctionalTest
into a new abstract class so that it could be shared with the new
PublicApiFunctionalTest.  There could be even more sharing, but it's
not entirely clear it'd be worthwhile at this point.  The new test
covers most features of the new public API.

Change-Id: I8c6859411f44b6430f8b348cf26a61225f5768cb