Add test for many interruptions to a single download.
Adding a new test case for downloads that undergo many interruptions
(as may happen with a very large download that takes many hours).
Includes some refactoring in the test suite.

Early on, this test exposed a race condition in which the download
manager got some I/O exception while reading from the MockWebServer.
I went in and improved/refactored much of the error logging code in
DownloadThread to try and track this down.  Unfortunately, once I
finished, the race condition no longer seems to be reproducible, even
with hundreds of runs of the test case.  So I've given up on it for
now.  In any event, error logging is better and much duplicate code
has been eliminated.