First pass at new system UI for download manager.
This is the initial checkin for a new system UI for the download
manager.  This UI shows all entries present in the Download Manager
database, sorted by either date or size.  It allows the user to
monitor status, open files, and delete/cancel entries.

The code was initially copied from the "Downloads" page of
packages/apps/Browser, since the design is based on that UI.  Much of
the code is changed, although DateSortedExpandableListAdapter is
nearly identical.

The code is not quite finished, but this is very much a working first
pass.

Change-Id: I7fe102e9f2998e57d893cc3c0f5f6da1cf2db549