(GB MR) bug:3144642 temporary small fix in GB MR and real fix in HC
Real fix is in Change-Id: Ifea1544737023008eff44aef9acd976902a0c143
In the database, sometimes _data column in downloads is set to null
and sometimes to empty string. this is inconsistent
and causes bugs such as bug:3144642.
This bug is caused by line# 793 in DownloadThread.
state.mFileName is null sometimes and empty string sometimes - because
the correspodning field is set inconsistentlt in downloads.db
_data column.

in GB MR, apply a bandaid because real fix could be too risky for
GB.

Change-Id: I115b3ba6cfe0262269dbbc4c336b9e1d63c618e4