fix broken DownloadManager tests
one big change in this CL is addition of a new feature to MockWebServer.
It can now play a long response to the Downloading thread to keep it busy
while something - such as cancel/remove - can be done to that Download Request.

Also, added changes to FakeSystemFacade to start threads in normal fashion
instead of queuing them up and later running just their run() methods.

the following tests should work now
	packages/providers/DownloadProvider/tests/src/com/android/providers/downloads/
	  DownloadManagerFunctionalTest.java
	  PublicApiFunctionalTest.java
	  ThreadingTest.java
	  PublicApiAccessTest.java
	  DownloadProviderPermissionsTest.java

the following are flaky. they need to be split up into smaller tests.
	frameworks/base/core/tests/coretests/src/android/app/
	  DownloadManagerIntegrationTest.java
	  DownloadManagerStressTest.java

Change-Id: Ia0b11963f92e8f8365f701761dcbce467be3ee9b