when download-complete-notif is clicked on, don't show notif anymore
bug:3272589
in HC, when using public API, if the downloaded file is in downloads data
dir, wrong intent action is set when the user clicks on download-complete
notification.
fix this by sending the correct intent action if the download is
in downloads data dir or any other location specified by the app.

Change-Id: I72621be5f1d6a5caecb8ec85634b351e08b66c31