(GB MR) bug:3351783 don't store redirectcount in db
redirectcount shouldn't be stored in the db.
say, you are downloading a large app and there are redirects
involved inthe download.
if network connectivity changes during download, then download
is paused and resumed once connectivy is back on.
when the download is resumed, it should start with redirectcount of 0
instead of redirectcount from the previous download session.

Change-Id: I1acf2fb9491d24f5e109206a91dda3eea429e447