Don't change extensions for explicitly set download locations.
This regression from GB was introduced by 38f1711910715538d6b64e505fe8fb9a0cd40475,
which was intended to allow duplicate downloads of the same file, adding -<n> to
the end of file names.  As a side effect, this also activated extension validation
logic, which adds/changes an extension to match the Mimetype.

This change keeps the unique filename logic but prevents extension changes when an
explicit filename is set.  Thus, it is still possible for the actual download
location to differ from the requested one, but only if the file already exists.

Bug: 5196436
Change-Id: I198dc2a819c5d839a05b72c25e0830d889a9c5a3