Deleting downloads for removed uids on downloadprovider start
After uninstalling an app, if the system was shutdown before the
download provider received the broadcast for UID_REMOVED, another app
installed later in the same uid might be able to gain access to the
files downloaded by this app. Removing any such hanging downloads
at the start up of the download provider should fix this issue.

Test: Manually tested by uninstalling an app and killing and restarting
the process android.process.media, to check that the downloaded files of
the uninstalled app were deleted.

Bug:22011579

Merged-In: I7382c4846f99035b40412a01715aee5873efa9e6
Change-Id: I7382c4846f99035b40412a01715aee5873efa9e6