Remove unnecessary direct dependency on dexmaker
The mockito-target library already statically includes dexmaker
so the direct dependency is unnecessary and could make the
upgrade of mockito more difficult.

This is part of the cleanup work prior to upgrading mockito.

I ran the tests before and after the change and in both cases
got 2 failures and between 4-6 errors; at least some of the
errors are flaky as the exact number changed from one run to the
next.

Bug: 32912773
Test: clean make of DownloadProviderTests, verify still includes dexmaker classes
Change-Id: I8133d0869a2b0ff7a9ffc19daff61ed1bbc6a7f6