Fix DownloadStorageProvider watching files in known download dirs.
This change will ensure when the download root is queried for
the first time, DownloadStorageProvider will start observing
the existing download dirs. There is still a corner case which
is not fixed by this change that is if a new sandbox top-level
Download directory gets created while the user is using the
files app, changes in that newly created directory won't be
picked up automatically unless the user refreshes
the app.

Bug: 124524422
Test: manual
Change-Id: I8d57eeb3f630039d217c4f0c8802c1b1b80da745