Set FGS priority for download jobs with visible notifications.
This is what DownloadProvider used to do pre-Q using
PRIORITY_FOREGROUND_APP but in Q, we added a new priority to denote
FGS but DownloadProvider was still using PRIORITY_FOREGROUND_APP which
is now considered a lower priority than FGS and some jobs are getting
preempted unnecessarily. So, update DownloadProvider to mark download
jobs with FGS priority instead of BFGS priority. This still has some
issues but at least this will be kind of reverting DownloadProvider
to its pre-Q behavior.

Bug: 135982638
Test: manual - Go through DO provisioning a few times and check admin
      app can be successfully downloaded.
Test: atest DownloadProviderTests
Test: atest cts/tests/app/src/android/app/cts/DownloadManagerTest.java
Test: atest cts/tests/app/DownloadManagerLegacyTest/src/android/app/cts/DownloadManagerLegacyTest.java
Test: atest cts/tests/app/DownloadManagerApi28Test/src/android/app/cts/DownloadManagerApi28Test.java
Change-Id: I020e366845d768da271b860a7ace91318e67d8ae