Progress towards splitting MediaProvider UID.
We're working towards making MediaProvider into a Mainline module,
which means it can no longer share the "android.media" UID built
into the OS.  The major reason for this is because they're signed
with separate keys, but a secondary reason is that we don't know what
database customizations an OEM has performed in their MediaProvider.

Enforce access based on permissions instead of UIDs.

Bug: 144247087
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I4b80456f0801b4a8a42fb442da3f40f34d0e4fa8