Give DownloadProvider the WRITE_MEDIA_STORAGE permission.
This is needed to get a full view of the filesystem on devices that
don't have scoped storage enabled yet. DownloadProvider requires a
writable view because it can write in app/OBB dirs on behalf of apps.
DownloadProvider already had this permission previously because it was
sharing a process with the MediaProvider, which requested that
permission; but now that MediaProvider has become a mainline module,
DownloadProvider no longer holds it.

Bug: 146378784
Test: N/A
Change-Id: I7f778ec13a335174d969a0bdec771d8b76c07cfb