Fixup files downloaded in Android/.
On devices without sdcardfs, paths in Android/data and Android/obb need
to be owned by the UID of the app. If DownloadProvider is asked to
create a file in one of these paths on the app's behalf, we need to
explicitly ask the system to fix up the permissions, to make sure they
are correct.

Bug: 152237221
Test: manually verified
Change-Id: Ife4df8c3ba322956a66db9b5c8718b93b2d0a325