Add CONNECTIVITY_USE_RESTRICTED_NETWORKS permission
- DownloadManager will use the requestor app's default network to
open a connection. When the VPN app lists the download provider
as a disallowed application, this will fail with EPERM.
- This is because in R, the download provider lost its privileges
due to the removal of the CONNECTIVITY_INTERNAL permission. It
was removed without considering the corner case where the
download provider is excluded from using the VPN. It's the only
case where this makes a difference.
- Adding CONNECTIVITY_USE_RESTRICTED_NETWORKS will give
download provider the ability to use the VPN again in this case,
fixing the issue.

Bug: 165774987
Test: Build, flash rom and boot to home. Manully test that
      DownloadProvider can download file via VPN.
Change-Id: I435655e4c8cff480342fa2a84b65a45a4194c4f3