Add HttpEngine to DownloadManager behind a flag
HttpEngine is the platform's current state-of-the-art HTTP stack that
is much more performant than URLConnection. The reasoning behind backing
DownloadManager with HttpEngine is a stepping-stone in making HttpEngine
the default stack for everything and slowly move away from URLConnection
which is outdated and is still using H/1.

At the moment, HttpEngine will only be enabled under two conditions:

1. Flag is enabled.
2. Caller APK does not have per-domain configs. This is necessary as
plumbing support for custom CAs / trust root is hard and doing it in
a single CL is going to be errorprone.

There's going to be a follow-up CL with plenty of metrics to evaluate
the impact of this flag.

Test: atest CtsNetSecConfigDownloadManagerTestCases (with flag enabled)
Test: atest DownloadProviderTests
Bug: 371965430
Change-Id: If91e3e4d61845cf276a0745a5d81aa540a74d4e1